Make TestimonialSubSection a reusable component
